+++ motors.h
@@ -39,8 +39,8 @@ class Motor{
 
     //these are timeout interupts and end distance values used to move the buggy a set distance
     Timeout *check_reached_distance_R, *check_reached_distance_L;
-    double end_distance_R, end_distance_L;
-     bool direction_R, direction_L;
+    long end_distance_R, end_distance_L;
+    bool direction_R, direction_L;
 
     //store the encoder counts and directions for encoders 
     int encoder_count_R, encoder_count_L;
@@ -63,8 +63,8 @@ class Motor{
     void update_speed_BB();     //bang bang
     void set_target_speed(double new_target_speed_R, double new_target_speed_L);    //set the target speed for the PID
     void move_constant_speed(double new_speed_R, double new_speed_L);     //move with no encoder feedback, arguments 0-1.0 sign gives direction
-    void move_distance_R(double distance, double speed);            //move the right wheel a certain distance
-    void move_distance_L(double distance, double speed);            //move the left wheel a certain distance
+    void move_distance_R(long distance, double speed);            //move the right wheel a certain distance
+    void move_distance_L(long distance, double speed);            //move the left wheel a certain distance
     void check_distance_R();
     void check_distance_L();
